Move `EmbeddedSchema` to its own file
This makes the code a bit easier to read. I’ve also
moved some of the tests over
