+++ lib/engines/content_block_manager/test/unit/app/models/content_block_schema_test.rb
@@ -62,132 +62,6 @@ class ContentBlockManager::SchemaTest < ActiveSupport::TestCase
         assert_equal subschemas.map(&:id), %w[bar]
       end
     end
-
-    describe "#subschema" do
-      it "returns a given subschema" do
-        subschema = schema.subschema("bar")
-
-        assert_equal subschema.id, "bar"
-        assert_equal subschema.fields, %w[my_string something_else]
-      end
-
-      describe "when an order is given in the subschema" do
-        let(:body) do
-          {
-            "properties" => {
-              "foo" => {
-                "type" => "string",
-              },
-              "bar" => {
-                "type" => "object",
-                "patternProperties" => {
-                  "*" => {
-                    "type" => "object",
-                    "order" => %w[something_else my_string],
-                    "properties" => {
-                      "my_string" => {
-                        "type" => "string",
-                      },
-                      "something_else" => {
-                        "type" => "string",
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          }
-        end
-
-        it "orders fields when an order is given" do
-          subschema = schema.subschema("bar")
-
-          assert_equal subschema.fields, %w[something_else my_string]
-        end
-      end
-
-      describe "when an order is given in the config" do
-        let(:body) do
-          {
-            "properties" => {
-              "rate" => {
-                "type" => "object",
-                "patternProperties" => {
-                  "*" => {
-                    "type" => "object",
-                    "properties" => {
-                      "name" => {
-                        "type" => "string",
-                      },
-                      "amount" => {
-                        "type" => "string",
-                      },
-                      "description" => {
-                        "type" => "string",
-                      },
-                      "frequency" => {
-                        "type" => "string",
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          }
-        end
-
-        before do
-          ContentBlockManager::ContentBlock::Schema
-            .stubs(:schema_settings)
-            .returns({
-              "schemas" => {
-                schema.id => {
-                  "subschemas" => {
-                    "rate" => {
-                      "field_order" => %w[name amount frequency description],
-                    },
-                  },
-                },
-              },
-            })
-        end
-
-        it "orders fields when an order is given" do
-          subschema = schema.subschema("rate")
-
-          assert_equal subschema.fields, %w[name amount frequency description]
-        end
-      end
-
-      describe "when an invalid subschema is given" do
-        let(:body) do
-          {
-            "properties" => {
-              "foo" => {
-                "type" => "string",
-              },
-              "bar" => {
-                "type" => "object",
-                "properties" => {
-                  "my_string" => {
-                    "type" => "string",
-                  },
-                  "something_else" => {
-                    "type" => "string",
-                  },
-                },
-              },
-            },
-          }
-        end
-
-        it "raises an error" do
-          assert_raises ArgumentError, "Subschema `bar` is invalid" do
-            schema.subschema("bar")
-          end
-        end
-      end
-    end
   end
 
   describe ".permitted_params" do
